Separate MCTS search from native recording
