+++ src/agents/mcts/mod.rs
@@ -1,4 +1,6 @@
 mod agent;
 mod node;
+mod search;
 
 pub use agent::MCTSAgent;
+pub use search::MCTSSearch;
