+++ src/agents/mod.rs
@@ -5,7 +5,7 @@ mod policy_sampling;
 mod random;
 
 pub use base::Agent;
-pub use mcts::MCTSAgent;
+pub use mcts::{MCTSAgent, MCTSSearch};
 pub use pentobi::PentobiAgent;
 pub use policy_sampling::PolicySamplingAgent;
 pub use random::RandomAgent;
